Use the EnvironmentData class to construct things
Updated EnvironmentModule to require the installation of a callback to the environment containing it as part of the constructor.

The create method is only for making the shared pointer version of this code. I am at the point though where I believe I will make Environment the name of the class itself and use Environment::Ptr everwhere else