Introduce the concept of an Entity to be the parent to the EntityRecord.
The entity record seems to be a way to apply the concept of virtual methods to C code. So I am wondering if it is possible to instead have all primitives be entities instead of having a separate record type for each...

Just food for thought. Then PatternEntityRecord is an extension of EntityRecord (PatternEntityRecord -> PatternEntity). Just an idea.

The idea of having install and deinstall methods still doesn't sit right with me in this case :/.

The getNext functions will have to go as well.