Thinking out aspects of the entity system
It looks like I can actually jettison the entity record as a separate class and just inherit from it in some fashion. Since we have atomic elements as well I'm not sure how to handle that either. The forms for both Atom and Entity must match properly.

Also drop the bitmap property because it isn't necessary anymore. We can use virtual methods to do the same thing.