Merge SymbolModule into Environment proper
When I did the Neutron library for Parasoft, I had an Environment wrapper class which wrapped around a CLIPS raw environment and added extra functionality. All of the structural additions (especially include) were saved within this wrapper environment class as fields.

There is no point in keeping the map of modules for the C++ version of clips. Adding new functionality is as simple as extending the environment class or just calling the associated methods provided by the environment class.

The added overhead of keeping things in separate objects can lead to potentially cleaner code but the need to keep track of the order of module creation is too much of a pain in the ass to keep.

This is the first "module" to be added to the environment class. The different structures that are part of these modules will stay just the state that is placed into the environment will be part of the environment itself.