ExternalAddress support
Support the new method via implementing a special ctor signature for a given ExternalAddress. If it isn't there then nullptr is returned and not registered in the factory tracker.

Support the optional call interface by implementing Callable as well as ExternalAddress. It will be detected as needed.

To prevent things from being deleted then it could be that we hold onto the thing via an observer_ptr but that is not implemented yet.